+++ complexity.py
@@ -89,22 +89,6 @@ def another_conv_block_complexity(model_config, input_shape):
     return cx, shape
 
 
-def res_basic_stage_complexity(model_config, input_shape):
-    # mandatory parameters
-    depth = model_config['depth']
-    strides = model_config['strides']
-
-    model_config = copy.deepcopy(model_config)
-    shape = input_shape
-    total_cx = {}
-
-    for i in range(depth):
-        cx, shape = res_basic_block_complexity(model_config, shape)
-        total_cx = dict_add(total_cx, cx)
-        model_config['strides'] = 1
-    return total_cx, shape
-
-
 def res_basic_block_complexity(model_config, input_shape):
     # mandatory parameters
     filters = model_config['filters']
@@ -127,22 +111,6 @@ def res_basic_block_complexity(model_config, input_shape):
                                   prev_cx=cx)
         cx, _ = norm_complexity(shape, prev_cx=cx)
     return cx, shape
-
-
-def res_bottleneck_stage_complexity(model_config, input_shape):
-    # mandatory parameters
-    depth = model_config['depth']
-    strides = model_config['strides']
-
-    model_config = copy.deepcopy(model_config)
-    shape = input_shape
-    total_cx = {}
-
-    for i in range(depth):
-        cx, shape = res_bottleneck_block_complexity(model_config, shape)
-        total_cx = dict_add(total_cx, cx)
-        model_config['strides'] = 1
-    return total_cx, shape
 
 
 def res_bottleneck_block_complexity(model_config, input_shape):
