+++ complexity_test.py
@@ -30,17 +30,6 @@ class ComplexityTest(tf.test.TestCase):
                              model_config,
                              [32, 32, 16])
 
-    def test_res_basic_stage_complexity(self):
-        model_config = {
-            'depth': 2,
-            'strides': 2,
-            'filters': 24,
-        }
-        self.complexity_test(res_basic_stage_complexity,
-                             res_basic_stage,
-                             model_config,
-                             [32, 32, 16])
-
     def test_res_basic_block_complexity(self):
         model_config = {
             'strides': 2,
@@ -50,19 +39,6 @@ class ComplexityTest(tf.test.TestCase):
                              res_basic_block,
                              model_config,
                              [32, 32, 3])
-
-    def test_res_bottleneck_stage_complexity(self):
-        model_config = {
-            'depth': 4,
-            'strides': 2,
-            'filters': 24,
-            'groups': 2,
-            'bottleneck_ratio': 2
-        }
-        self.complexity_test(res_bottleneck_stage_complexity,
-                             res_bottleneck_stage,
-                             model_config,
-                             [32, 32, 16])
 
     def test_res_bottleneck_block_complexity(self):
         model_config = {
